+++ json_serializable/lib/builder.dart
@@ -27,7 +27,8 @@ Builder jsonSerializable(BuilderOptions options) {
 
   var builder = jsonPartBuilder(
       header: optionsMap.remove('header') as String,
-      useWrappers: optionsMap.remove('use_wrappers') as bool);
+      useWrappers: optionsMap.remove('use_wrappers') as bool,
+      anyMap: optionsMap.remove('any_map') as bool);
 
   if (optionsMap.isNotEmpty) {
     if (log == null) {