+++ json_serializable/lib/src/type_helpers/convert_helper.dart
@@ -30,34 +30,10 @@ class ConvertHelper extends TypeHelper {
       DartType targetType, String expression, DeserializeContext context) {
     var fromJsonData = (context as TypeHelperContext).fromJsonData;
     if (fromJsonData != null) {
-      var asContent = _asContent(fromJsonData.paramType);
+      var asContent = asStatement(fromJsonData.paramType);
       var result = '${fromJsonData.name}($expression$asContent)';
       return commonNullPrefix(context.nullable, expression, result);
     }
     return null;
   }
-}
-
-String _asContent(DartType type) {
-  if (type.isDynamic || type.isObject) {
-    return '';
-  }
-
-  if (coreIterableTypeChecker.isAssignableFromType(type)) {
-    var itemType = coreIterableGenericType(type);
-    if (itemType.isDynamic || itemType.isObject) {
-      return ' as List';
-    }
-  }
-
-  if (coreMapTypeChecker.isAssignableFromType(type)) {
-    var args = typeArgumentsOf(type, coreMapTypeChecker);
-    assert(args.length == 2);
-
-    if (args.every((dt) => dt.isDynamic || dt.isObject)) {
-      return ' as Map';
-    }
-  }
-
-  return ' as $type';
 }