+++ json_serializable/test/test_files/kitchen_sink.non_nullable.wrapped.dart
@@ -18,11 +18,13 @@
 import 'package:json_annotation/json_annotation.dart';
 
 import 'kitchen_sink_interface.dart' as k;
+import 'simple_object.dart';
 
 part 'kitchen_sink.non_nullable.wrapped.g.dart';
 
 List<T> _defaultList<T>() => <T>[];
 Map<String, T> _defaultMap<T>() => <String, T>{};
+SimpleObject _defaultSimpleObject() => new SimpleObject(42);
 
 k.KitchenSink testFactory(
         {int ctorValidatedNo42,
@@ -39,8 +41,7 @@ k.KitchenSink testFactory(
         intIterable: intIterable,
         dateTimeIterable: dateTimeIterable);
 
-k.KitchenSink testFromJson(Map<String, dynamic> json) =>
-    new KitchenSink.fromJson(json);
+k.KitchenSink testFromJson(Map json) => new KitchenSink.fromJson(json);
 
 @JsonSerializable(nullable: false)
 class KitchenSink extends Object
@@ -80,8 +81,7 @@ class KitchenSink extends Object
     }
   }
 
-  factory KitchenSink.fromJson(Map<String, Object> json) =>
-      _$KitchenSinkFromJson(json);
+  factory KitchenSink.fromJson(Map json) => _$KitchenSinkFromJson(json);
 
   @JsonKey(includeIfNull: false)
   DateTime dateTime = new DateTime(1981, 6, 5);
@@ -118,6 +118,8 @@ class KitchenSink extends Object
   bool writeNotNull;
   @JsonKey(name: r'$string')
   String string;
+
+  SimpleObject simpleObject = _defaultSimpleObject();
 
   bool operator ==(Object other) => k.sinkEquals(this, other);
 }