+++ json_serializable/test/test_files/kitchen_sink.non_nullable.wrapped.g.dart
@@ -10,7 +10,7 @@ part of 'kitchen_sink.non_nullable.wrapped.dart';
 // Generator: JsonSerializableGenerator
 // **************************************************************************
 
-KitchenSink _$KitchenSinkFromJson(Map<String, dynamic> json) => new KitchenSink(
+KitchenSink _$KitchenSinkFromJson(Map json) => new KitchenSink(
     ctorValidatedNo42: json['no-42'] as int,
     iterable: json['iterable'] as List,
     dynamicIterable: json['dynamicIterable'] as List,
@@ -26,24 +26,25 @@ KitchenSink _$KitchenSinkFromJson(Map<String, dynamic> json) => new KitchenSink(
   ..dateTimeList = (json['dateTimeList'] as List)
       .map((e) => DateTime.parse(e as String))
       .toList()
-  ..map = json['map'] as Map<String, dynamic>
+  ..map = json['map'] as Map
   ..stringStringMap =
       new Map<String, String>.from(json['stringStringMap'] as Map)
   ..stringIntMap = new Map<String, int>.from(json['stringIntMap'] as Map)
   ..stringDateTimeMap = new Map<String, DateTime>.fromIterables(
-      (json['stringDateTimeMap'] as Map<String, dynamic>).keys,
+      (json['stringDateTimeMap'] as Map).keys.cast<String>(),
       (json['stringDateTimeMap'] as Map)
           .values
           .map((e) => DateTime.parse(e as String)))
   ..crazyComplex = (json['crazyComplex'] as List)
       .map((e) => new Map<String, Map<String, List<List<DateTime>>>>.fromIterables(
-          (e as Map<String, dynamic>).keys,
-          (e as Map).values.map((e) =>
-              new Map<String, List<List<DateTime>>>.fromIterables((e as Map<String, dynamic>).keys, (e as Map).values.map((e) => (e as List).map((e) => (e as List).map((e) => DateTime.parse(e as String)).toList()).toList())))))
+          (e as Map).keys.cast<String>(),
+          (e as Map).values.map(
+              (e) => new Map<String, List<List<DateTime>>>.fromIterables((e as Map).keys.cast<String>(), (e as Map).values.map((e) => (e as List).map((e) => (e as List).map((e) => DateTime.parse(e as String)).toList()).toList())))))
       .toList()
   ..val = new Map<String, bool>.from(json['val'] as Map)
   ..writeNotNull = json['writeNotNull'] as bool
-  ..string = json[r'$string'] as String;
+  ..string = json[r'$string'] as String
+  ..simpleObject = new SimpleObject.fromJson(json['simpleObject'] as Map);
 
 abstract class _$KitchenSinkSerializerMixin {
   int get ctorValidatedNo42;
@@ -66,6 +67,7 @@ abstract class _$KitchenSinkSerializerMixin {
   Map<String, bool> get val;
   bool get writeNotNull;
   String get string;
+  SimpleObject get simpleObject;
   Map<String, dynamic> toJson() => new _$KitchenSinkJsonMapWrapper(this);
 }
 
@@ -94,7 +96,8 @@ class _$KitchenSinkJsonMapWrapper extends $JsonMapWrapper {
         'crazyComplex',
         'val',
         'writeNotNull',
-        r'$string'
+        r'$string',
+        'simpleObject'
       ];
 
   @override
@@ -152,6 +155,8 @@ class _$KitchenSinkJsonMapWrapper extends $JsonMapWrapper {
           return _v.writeNotNull;
         case r'$string':
           return _v.string;
+        case 'simpleObject':
+          return _v.simpleObject;
       }
     }
     return null;