+++ json_serializable/tool/build.dart
@@ -31,16 +31,28 @@ final List<BuilderApplication> builders = [
           'example/example.dart',
           'test/test_files/json_literal.dart',
           'test/test_files/json_test_example.dart',
-          'test/test_files/json_test_example.non_nullable.dart',
+          'test/test_files/json_test_example.non_nullable.dart'
+        ],
+      )),
+  applyToRoot(jsonPartBuilder(header: copyrightHeader, anyMap: true),
+      generateFor: const InputSet(
+        include: const [
           'test/test_files/kitchen_sink.dart',
-          'test/test_files/kitchen_sink.non_nullable.dart'
+          'test/test_files/kitchen_sink.non_nullable.dart',
+          'test/test_files/simple_object.dart'
         ],
       )),
   applyToRoot(jsonPartBuilder(useWrappers: true, header: copyrightHeader),
       generateFor: const InputSet(
         include: const [
-          'test/test_files/kitchen_sink*wrapped.dart',
           'test/test_files/json_test_example*wrapped.dart',
+        ],
+      )),
+  applyToRoot(
+      jsonPartBuilder(useWrappers: true, anyMap: true, header: copyrightHeader),
+      generateFor: const InputSet(
+        include: const [
+          'test/test_files/kitchen_sink*wrapped.dart',
         ],
       )),
   applyToRoot(testBootstrapBuilder(null),